Moved single_city_page to pages/singlecity
