+++ src/components/pages/singlecity/SCPactivities.js
