+++ src/components/pages/singlecity/SCPfashion.js
