+++ src/components/pages/singlecity/SCPfurniture.js
