+++ src/components/pages/singlecity/SCPhardware.js
