+++ src/components/pages/singlecity/SCPhotels.js
