+++ src/components/pages/singlecity/SingleCityPage.scss
